+++ python_bindings/src/PyIROperator.cpp
@@ -109,18 +109,7 @@ void define_operators(py::module &m) {
             << "tuple_select() may not mix Expr and Tuple for the condition elements.";
         return to_python_tuple(false_value);
     });
-
-    m.def("select_by_id", [](py::args args) -> Expr {
-        if (args.size() < 2) {
-            throw py::value_error("select_by_id() must have at least 2 arguments");
-        }
-        std::vector<Expr> values;
-        values.reserve(args.size());
-        for (int i = 1; i < (int)args.size(); i++) {
-            values.push_back(args[i].cast<Expr>());
-        }
-        return select_by_id(args[0].cast<Expr>(), values);
-    });
+    m.def("select_by_index", &select_by_index);
 
     m.def("sin", &sin);
     m.def("asin", &asin);
